sscs 8081 moved to new content and fixed tests
